Make some metal functions callable outside of metal return.c
